+++ src/test/java/racing/model/PlayHistoryTest.java
@@ -1,10 +1,12 @@
-package racing;
+package racing.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.model.CarPositions;
+import racing.model.PlayHistory;
 
 public class PlayHistoryTest {
