+++ src/test/java/racing/model/RacingSessionTest.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -6,6 +6,9 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import java.util.List;
+import racing.model.Car;
+import racing.model.CarPositions;
+import racing.model.RacingSession;
 
 import static org.assertj.core.api.Assertions.assertThat;
